Add checkbox toggle to update to-do list item's `done` status
The checkbox toggle updates and saves the to-do's `done` property to `true` or `false`.

`onCheckboxToggle` runs when a user clicks on a to-do list item's checkbox. Once a checkbox is checked, the to-do's `done` property is set to `true` and its text is strikedthrough. When it is unchecked, its value returns to its default `false` and the text has no decoration.

Data persists after sign out.